backend/requirements.txt
# Core LLM and Agent Framework
langchain==0.3.27
langchain-groq==0.3.8
langchain-community==0.3.31
langchain-text-splitters==0.3.11

# Vector Store and Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
pinecone-client==6.0.0
langchain-pinecone==0.2.13

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
unstructured==0.12.0

# Search and Retrieval
rank-bm25==0.2.2

# API and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# UI
streamlit==1.31.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
numpy==1.26.3
loguru==0.7.2

# Testing
pytest==8.0.0